Bug #31326120 UBSAN: SHIFT EXPONENT 4294967294 IS TOO LARGE FOR 64-BIT TYPE
The basic lookup function find_type() in typelib returns an int -1 in
case of ambiguities.  When this is converted to unsigned, and used as
an argument for left-shift, we get undefined behaviour.

Propagate the -1, and handle it as an error before doing any left-shifting.

+++ include/typelib.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2000, 2019, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2000, 2020, Oracle and/or its affiliates.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License, version 2.0,
@@ -58,7 +58,7 @@ extern TYPELIB *copy_typelib(MEM_ROOT *root, TYPELIB *from);
 
 extern TYPELIB sql_protocol_typelib;
 
-uint64_t find_set_from_flags(const TYPELIB *lib, size_t default_name,
+uint64_t find_set_from_flags(const TYPELIB *lib, int default_name,
                              uint64_t cur_set, uint64_t default_set,
                              const char *str, unsigned int length,
                              const char **err_pos, unsigned int *err_len);
